update smallint to integer

+++ db/schema/06_users.sql
@@ -7,7 +7,7 @@ CREATE TABLE users (
   password        VARCHAR(255) NOT NULL,
   email           VARCHAR(255),
   nickname        VARCHAR(255),
-  phone           SMALLINT,
+  phone           INTEGER,
   profile_avatar  TEXT,
   is_teacher      BOOLEAN DEFAULT false,
   created_on      DATE DEFAULT NOW(),
